# Remove redundant key encoding & make `Transaction` dyn compatible.