--- a/crates/core/src/doc/upsert.rs
+++ b/crates/core/src/doc/upsert.rs
@@ -25,7 +25,7 @@ impl Document {
 			return self.upsert_update(stk, ctx, opt, stm).await;
 		}
 
-		ctx.tx().lock().await.new_save_point().await;
+		ctx.tx().lock().await.new_save_point();
 
 		// First try to create the value and if that is not possible due to an existing value fall
 		// back to update instead.
@@ -64,11 +64,11 @@ impl Document {
 				}
 			},
 			Err(IgnoreError::Ignore) => {
-				ctx.tx().lock().await.release_last_save_point().await?;
+				ctx.tx().lock().await.release_last_save_point()?;
 				return Err(IgnoreError::Ignore);
 			}
 			Ok(x) => {
-				ctx.tx().lock().await.release_last_save_point().await?;
+				ctx.tx().lock().await.release_last_save_point()?;
 				return Ok(x);
 			}
 		};