--- a/crates/core/src/key/debug.rs
+++ b/crates/core/src/key/debug.rs
@@ -1,6 +1,10 @@
+//! Debugging utilities for the `key` crate.
+
 use std::ops::Range;
 
+/// A trait for types that can be converted to a string representation for readability.
 pub trait Sprintable {
+	/// Converts the implementing type to a string representation.
 	fn sprint(&self) -> String;
 }