--- a/crates/core/src/kvs/ds.rs
+++ b/crates/core/src/kvs/ds.rs
@@ -133,32 +133,32 @@ impl TransactionFactory {
 			#[cfg(feature = "kv-mem")]
 			DatastoreFlavor::Mem(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, true, false)
+				(tx, true, false)
 			}
 			#[cfg(feature = "kv-rocksdb")]
 			DatastoreFlavor::RocksDB(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, true, true)
+				(tx, true, true)
 			}
 			#[cfg(feature = "kv-indxdb")]
 			DatastoreFlavor::IndxDB(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, true, false)
+				(tx, true, false)
 			}
 			#[cfg(feature = "kv-tikv")]
 			DatastoreFlavor::TiKV(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, false, true)
+				(tx, false, true)
 			}
 			#[cfg(feature = "kv-fdb")]
 			DatastoreFlavor::FoundationDB(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, false, false)
+				(tx, false, false)
 			}
 			#[cfg(feature = "kv-surrealkv")]
 			DatastoreFlavor::SurrealKV(v) => {
 				let tx = v.transaction(write, lock).await?;
-				(Box::new(tx) as Box<dyn super::api::Transaction>, true, false)
+				(tx, true, false)
 			}
 			_ => unreachable!(),
 		};