Thank you for submitting this pull request. We really appreciate you spending the time to work on SurrealDB. 🚀 🎉 

## What is the motivation?

<!-- Please provide details on the motivation for why you have made this change.-->

We have multiple layers of transactions:
`struct Transaction` -> `struct Transactor` -> `enum Inner` -> `struct db_specific::Transaction`.

At all layers of this transaction hierarchy, we're calling `encode_owned()` on the incoming keys. This results in, at best, a `Vec<u8>` getting moved many times:
```rust
impl KeyEncode for Vec<u8> {
	fn encode(&self) -> Result<Vec<u8>, crate::err::Error> {
		Ok(self.clone())
	}

    fn encode_owned(self) -> Result<Vec<u8>, crate::err::Error> {
		Ok(self)
	}
}
```
and at worst a `&[u8]` getting cloned into a `Vec<u8>` potentially many times:
```rust
impl KeyEncode for &[u8] {
    fn encode_owned(self) -> Result<Vec<u8>, crate::err::Error>
	where
		Self: Sized,
	{
		self.encode()
	}

	fn encode(&self) -> Result<Vec<u8>, crate::err::Error> {
		let mut buf = Vec::new();
		self.encode_into(&mut buf)?;
		Ok(buf)
	}

	fn encode_into(&self, buffer: &mut Vec<u8>) -> Result<(), crate::err::Error> {
		buffer.extend_from_slice(self);
		Ok(())
	}
}
```

## What does this change do?

<!-- Please provide a description of what this pull request does, and how it solves the problem. -->

This change modifies the `trait Transaction` to be dyn compatible and removes the `tr::Inner` enum. In order for the Transaction to be dyn compatible, by removing the generics (eg `K: KeyEncode`). All key encodings occur at the `Transactor` layer now.

## What is your testing strategy?

<!-- Write your test plan here. Please provide us with clear instructions on how you verified your changes work. -->

Passing CI tests should be sufficient.

### Performance verification

Overall, as expected, this change makes little difference in terms of performance. Getting a transaction is not our performance bottleneck, so the `crud-bench` results are all on the margins. Below is a comparison between main and this branch.

These results are all from running on my M4 mac.

#### Main

Server:
```
$ git rev-parse --short HEAD
50df71c5
$ cargo run --release --no-default-features --features storage-mem,http,scripting -- start --user root --pass root memory
```

```
$ cargo run -r -- -d surrealdb -e ws://127.0.0.1:8000 -s 1000000 -c 12 -t 24 -r
--------------------------------------------------
Create 100% - Create took 17s 781ms
Read 100% - Read took 2s 500ms
Update 100% - Update took 19s 366ms
Scan::count_all 100% - Scan::count_all took 6s 285ms
Scan::limit_id 100% - Scan::limit_id took 11ms 50µs
Scan::limit_all 100% - Scan::limit_all took 10ms 960µs
Scan::limit_count 100% - Scan::limit_count took 14ms 469µs
Scan::limit_start_id 100% - Scan::limit_start_id took 25ms 525µs
Scan::limit_start_all 100% - Scan::limit_start_all took 26ms 432µs
Scan::limit_start_count 100% - Scan::limit_start_count took 27ms 544µs
Delete 100% - Delete took 18s 411ms
--------------------------------------------------
Benchmark result for Surrealdb; endpoint => ws://127.0.0.1:8000
CPUs: 16 - Blocking threads: 16 - Workers: 16 - Clients: 12 - Threads: 24 - Samples: 1000000 - Key: Integer - Random: true
--------------------------------------------------
╭─────────────────────────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬─────────┬───────────┬────────┬──────────┬───────┬────────┬───────────────────╮
│ Test                            ┆ Total time ┆ Mean       ┆ Max        ┆ 99th       ┆ 95th       ┆ 75th       ┆ 50th       ┆ 25th       ┆ 1st        ┆ Min        ┆ IQR     ┆ OPS       ┆    CPU ┆   Memory ┆ Reads ┆ Writes ┆ System load       │
╞═════════════════════════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪═════════╪═══════════╪════════╪══════════╪═══════╪════════╪═══════════════════╡
│ [C]reate                        ┆ 17s 781ms  ┆ 5.12 ms    ┆ 24.32 ms   ┆ 6.93 ms    ┆ 6.52 ms    ┆ 5.80 ms    ┆ 5.23 ms    ┆ 4.53 ms    ┆ 2.53 ms    ┆ 0.32 ms    ┆ 1.27 ms ┆ 56239.30  ┆  7.09% ┆ 54.3 MiB ┆   0 B ┆    0 B ┆ 18.16/18.11/18.10 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [R]ead                          ┆ 2s 500ms   ┆ 0.72 ms    ┆ 8.46 ms    ┆ 1.81 ms    ┆ 1.24 ms    ┆ 0.83 ms    ┆ 0.66 ms    ┆ 0.56 ms    ┆ 0.14 ms    ┆ 0.05 ms    ┆ 0.27 ms ┆ 399997.77 ┆ 19.98% ┆ 64.6 MiB ┆   0 B ┆    0 B ┆ 19.35/18.36/18.19 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [U]pdate                        ┆ 19s 366ms  ┆ 5.58 ms    ┆ 24.64 ms   ┆ 7.04 ms    ┆ 6.67 ms    ┆ 6.10 ms    ┆ 5.65 ms    ┆ 5.13 ms    ┆ 3.46 ms    ┆ 0.30 ms    ┆ 0.97 ms ┆ 51634.38  ┆  6.55% ┆ 79.0 MiB ┆   0 B ┆    0 B ┆ 20.70/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::count_all (100)         ┆ 6s 285ms   ┆ 6075.86 ms ┆ 6287.36 ms ┆ 6287.36 ms ┆ 6287.36 ms ┆ 6287.36 ms ┆ 6287.36 ms ┆ 6287.36 ms ┆ 1049.60 ms ┆ 1048.58 ms ┆ 0.00 ms ┆ 15.91     ┆  0.01% ┆ 84.7 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_id (100)          ┆ 11ms 50µs  ┆ 7.57 ms    ┆ 10.86 ms   ┆ 10.83 ms   ┆ 10.56 ms   ┆ 8.81 ms    ┆ 7.49 ms    ┆ 6.60 ms    ┆ 1.48 ms    ┆ 1.48 ms    ┆ 2.20 ms ┆ 9049.70   ┆  0.01% ┆ 86.8 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_all (100)         ┆ 10ms 960µs ┆ 7.50 ms    ┆ 10.75 ms   ┆ 10.73 ms   ┆ 9.99 ms    ┆ 8.82 ms    ┆ 7.64 ms    ┆ 6.32 ms    ┆ 0.73 ms    ┆ 0.73 ms    ┆ 2.50 ms ┆ 9123.84   ┆  0.01% ┆ 89.0 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_count (100)       ┆ 14ms 469µs ┆ 8.77 ms    ┆ 14.13 ms   ┆ 14.10 ms   ┆ 13.10 ms   ┆ 10.54 ms   ┆ 8.64 ms    ┆ 6.67 ms    ┆ 2.50 ms    ┆ 2.50 ms    ┆ 3.87 ms ┆ 6910.97   ┆  0.00% ┆ 90.0 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_id (100)    ┆ 25ms 525µs ┆ 21.05 ms   ┆ 25.28 ms   ┆ 25.21 ms   ┆ 24.75 ms   ┆ 23.20 ms   ┆ 21.41 ms   ┆ 19.87 ms   ┆ 4.59 ms    ┆ 4.58 ms    ┆ 3.33 ms ┆ 3917.70   ┆  0.01% ┆ 90.8 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_all (100)   ┆ 26ms 432µs ┆ 21.21 ms   ┆ 26.09 ms   ┆ 26.06 ms   ┆ 25.18 ms   ┆ 23.52 ms   ┆ 21.87 ms   ┆ 19.70 ms   ┆ 4.39 ms    ┆ 4.39 ms    ┆ 3.82 ms ┆ 3783.29   ┆  0.01% ┆ 91.8 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_count (100) ┆ 27ms 544µs ┆ 22.03 ms   ┆ 27.31 ms   ┆ 27.28 ms   ┆ 26.61 ms   ┆ 24.66 ms   ┆ 22.54 ms   ┆ 20.30 ms   ┆ 4.52 ms    ┆ 4.52 ms    ┆ 4.35 ms ┆ 3630.46   ┆  0.00% ┆ 93.0 MiB ┆   0 B ┆    0 B ┆ 20.57/18.74/18.33 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [D]elete                        ┆ 18s 411ms  ┆ 5.30 ms    ┆ 31.73 ms   ┆ 7.42 ms    ┆ 6.62 ms    ┆ 5.90 ms    ┆ 5.37 ms    ┆ 4.75 ms    ┆ 3.06 ms    ┆ 0.29 ms    ┆ 1.14 ms ┆ 54314.39  ┆  6.66% ┆ 97.9 MiB ┆   0 B ┆    0 B ┆ 21.69/19.12/18.48 │
╰─────────────────────────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴─────────┴───────────┴────────┴──────────┴───────┴────────┴───────────────────╯
--------------------------------------------------
```

### This change
Server:
```rust
$ git rev-parse --short HEAD
fd17bdbb
$ cargo run --release --no-default-features --features storage-mem,http,scripting -- start --user root --pass root memory
```

```
$ cargo run -r -- -d surrealdb -e ws://127.0.0.1:8000 -s 1000000 -c 12 -t 24 -r
--------------------------------------------------
Create 100% - Create took 17s 508ms
Read 100% - Read took 2s 320ms
Update 100% - Update took 19s 308ms
Scan::count_all 100% - Scan::count_all took 5s 478ms
Scan::limit_id 100% - Scan::limit_id took 12ms 396µs
Scan::limit_all 100% - Scan::limit_all took 12ms 770µs
Scan::limit_count 100% - Scan::limit_count took 14ms 519µs
Scan::limit_start_id 100% - Scan::limit_start_id took 27ms 49µs
Scan::limit_start_all 100% - Scan::limit_start_all took 28ms 392µs
Scan::limit_start_count 100% - Scan::limit_start_count took 28ms 765µs
Delete 100% - Delete took 18s 61ms
--------------------------------------------------
Benchmark result for Surrealdb; endpoint => ws://127.0.0.1:8000
CPUs: 16 - Blocking threads: 16 - Workers: 16 - Clients: 12 - Threads: 24 - Samples: 1000000 - Key: Integer - Random: true
--------------------------------------------------
╭─────────────────────────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬────────────┬───────────┬───────────┬─────────┬───────────┬────────┬──────────┬───────┬────────┬───────────────────╮
│ Test                            ┆ Total time ┆ Mean       ┆ Max        ┆ 99th       ┆ 95th       ┆ 75th       ┆ 50th       ┆ 25th       ┆ 1st       ┆ Min       ┆ IQR     ┆ OPS       ┆    CPU ┆   Memory ┆ Reads ┆ Writes ┆ System load       │
╞═════════════════════════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪════════════╪═══════════╪═══════════╪═════════╪═══════════╪════════╪══════════╪═══════╪════════╪═══════════════════╡
│ [C]reate                        ┆ 17s 508ms  ┆ 5.04 ms    ┆ 18.70 ms   ┆ 6.79 ms    ┆ 6.39 ms    ┆ 5.70 ms    ┆ 5.15 ms    ┆ 4.48 ms    ┆ 2.67 ms   ┆ 0.33 ms   ┆ 1.22 ms ┆ 57115.24  ┆  7.18% ┆ 53.4 MiB ┆   0 B ┆    0 B ┆ 17.23/16.74/16.34 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [R]ead                          ┆ 2s 320ms   ┆ 0.67 ms    ┆ 11.46 ms   ┆ 1.71 ms    ┆ 1.17 ms    ┆ 0.77 ms    ┆ 0.60 ms    ┆ 0.51 ms    ┆ 0.14 ms   ┆ 0.04 ms   ┆ 0.26 ms ┆ 431027.31 ┆ 20.01% ┆ 66.7 MiB ┆   0 B ┆    0 B ┆ 17.23/16.74/16.34 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [U]pdate                        ┆ 19s 308ms  ┆ 5.56 ms    ┆ 35.90 ms   ┆ 7.03 ms    ┆ 6.64 ms    ┆ 6.08 ms    ┆ 5.63 ms    ┆ 5.12 ms    ┆ 3.50 ms   ┆ 0.29 ms   ┆ 0.96 ms ┆ 51790.96  ┆  6.60% ┆ 81.9 MiB ┆   0 B ┆    0 B ┆ 19.27/17.24/16.53 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::count_all (100)         ┆ 5s 478ms   ┆ 5295.69 ms ┆ 5480.45 ms ┆ 5480.45 ms ┆ 5480.45 ms ┆ 5480.45 ms ┆ 5480.45 ms ┆ 5480.45 ms ┆ 910.85 ms ┆ 910.34 ms ┆ 0.00 ms ┆ 18.25     ┆  0.01% ┆ 86.6 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_id (100)          ┆ 12ms 396µs ┆ 8.51 ms    ┆ 12.06 ms   ┆ 12.04 ms   ┆ 11.86 ms   ┆ 10.23 ms   ┆ 8.78 ms    ┆ 6.77 ms    ┆ 1.16 ms   ┆ 1.16 ms   ┆ 3.46 ms ┆ 8066.52   ┆  0.01% ┆ 89.6 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_all (100)         ┆ 12ms 770µs ┆ 8.82 ms    ┆ 12.57 ms   ┆ 12.57 ms   ┆ 12.18 ms   ┆ 10.71 ms   ┆ 8.64 ms    ┆ 7.19 ms    ┆ 1.32 ms   ┆ 1.32 ms   ┆ 3.52 ms ┆ 7830.62   ┆  0.01% ┆ 92.2 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_count (100)       ┆ 14ms 519µs ┆ 8.95 ms    ┆ 14.31 ms   ┆ 14.29 ms   ┆ 13.29 ms   ┆ 11.13 ms   ┆ 8.93 ms    ┆ 6.72 ms    ┆ 1.46 ms   ┆ 1.46 ms   ┆ 4.41 ms ┆ 6887.39   ┆  0.00% ┆ 93.6 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_id (100)    ┆ 27ms 49µs  ┆ 21.04 ms   ┆ 26.77 ms   ┆ 26.73 ms   ┆ 25.95 ms   ┆ 23.50 ms   ┆ 21.28 ms   ┆ 19.82 ms   ┆ 4.96 ms   ┆ 4.96 ms   ┆ 3.68 ms ┆ 3696.94   ┆  0.01% ┆ 94.0 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_all (100)   ┆ 28ms 392µs ┆ 22.18 ms   ┆ 28.08 ms   ┆ 28.05 ms   ┆ 27.15 ms   ┆ 25.17 ms   ┆ 22.69 ms   ┆ 20.32 ms   ┆ 4.33 ms   ┆ 4.32 ms   ┆ 4.85 ms ┆ 3522.04   ┆  0.01% ┆ 95.2 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [S]can::limit_start_count (100) ┆ 28ms 765µs ┆ 22.89 ms   ┆ 28.53 ms   ┆ 28.51 ms   ┆ 27.95 ms   ┆ 26.06 ms   ┆ 23.21 ms   ┆ 20.98 ms   ┆ 4.43 ms   ┆ 4.43 ms   ┆ 5.09 ms ┆ 3476.44   ┆  0.00% ┆ 96.2 MiB ┆   0 B ┆    0 B ┆ 19.25/17.27/16.54 │
├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
│ [D]elete                        ┆ 18s 61ms   ┆ 5.20 ms    ┆ 31.47 ms   ┆ 7.20 ms    ┆ 6.48 ms    ┆ 5.77 ms    ┆ 5.25 ms    ┆ 4.65 ms    ┆ 3.09 ms   ┆ 0.32 ms   ┆ 1.12 ms ┆ 55365.55  ┆  6.78% ┆ 99.7 MiB ┆   0 B ┆    0 B ┆ 20.63/17.71/16.72 │
╰─────────────────────────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴────────────┴───────────┴───────────┴─────────┴───────────┴────────┴──────────┴───────┴────────┴───────────────────╯
--------------------------------------------------
```

## Is this related to any issues?

<!-- If this pull request is related to other pull requests, or resolves any issues, then link all related or closed items here, using 'Closes #101' or 'Fixes #101' to automatically close any linked issues. -->

- [x] No related issues

## Does this change need documentation?

<!-- If this pull request requires changes, updates, or improvements to the documentation, then add a corresponding issue on the https://github.com/surrealdb/docs.surrealdb.com repository, and link to it here. -->

- [x] No documentation needed

## Have you read the Contributing Guidelines?

<!-- All pull requests require that the contributing guidelines have been read and agreed to. -->

- [x] I have read the [Contributing Guidelines](https://github.com/surrealdb/surrealdb/blob/main/CONTRIBUTING.md)